requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alabaster==0.7.16
attrs==23.2.0
Babel==2.15.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
docutils==0.21.2
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.7
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.4
libsql-client==0.3.1
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.0
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.0.0
snowballstemmer==2.2.0
Sphinx==7.3.7
sphinx-press-theme==0.8.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
SQLAlchemy==2.0.30
sqlalchemy-libsql==0.1.0
typing_extensions==4.12.0
urllib3==2.2.1
Werkzeug==2.0.3
yarl==1.9.4
